Arena 1 Finalization, Adding Darkness Effects
In the process of finalizing arenas. Arena 1 is finished, with darkness effects added. Increased size of Spider enemy and the spider projectile ("Sphere" in code), as well as the range from which the Spider enemy will fire at the player. Working on implementing same changes for other Arena scenes, but will leave Arena 2 unchanged for Game Manager HUD implementation.